+++ src/config.ts
@@ -1,0 +1,13 @@
+export type TConfig = {
+  simulateErrors?: boolean;
+  isWithoutTimeout?: boolean;
+};
+
+export const defaultConfig: TConfig = {
+  simulateErrors: false,
+  isWithoutTimeout: false
+};
+
+export class Config {
+  static joinedConfig: TConfig;
+}
